==> 4_ensemble.ipynb <==
import pandas as pd
import os
import numpy as np

file_list = os.listdir('./data/ensemble_submit/')
file_list_csv = [file for file in file_list if file.endswith(".csv") and file != 'ensemble_idsplit.csv']

sort_list = sorted(file_list_csv, reverse=True, key=lambda x: float(x.split('_')[-1].split('.csv')[0]))
sort_list

preds = []
for i in range(10):
    preds.append(pd.read_csv(sort_list[i]).iloc[:,1:])

preds = np.array(preds)
preds = np.mean(preds, axis=0)

preds.shape

ss = pd.read_csv('./data/sample_submission.csv')
ss.iloc[:,1:] = preds.astype(int)
ss.to_csv('./data/ensemble_submit/ensemble_idsplit.csv', index=False)
