==> 5_submit_after_process.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import os
from tqdm import tqdm

data = pd.read_csv('./data/train.csv')
submit_origin = pd.read_csv('./data/ensemble_submit/ensemble_idsplit.csv')
submit = submit_origin.copy()

all_data = data.iloc[:, -90:]
cols = submit.columns[1:]

for num in tqdm(range(10000,0, -1)):
    indices = all_data[(0 <= all_data) & (all_data <= num)].dropna().index

    submit.loc[indices, cols] = np.where(submit.loc[indices, cols] > num, num, submit.loc[indices, cols])

differences = (submit_origin.iloc[:, 1:] != submit.iloc[:, 1:])
diff_list = differences.any(axis=1)
diff_list = diff_list.index[diff_list].tolist()

len(diff_list) # -1 : 3530

# len(np.where(submit.iloc[:,1:] > 50000)[0])

# # submit.iloc[:,1:]의 데이터 중 50000 이상인 데이터는 모두 50000으로 바꾼다.
# submit.iloc[:,1:] = np.where(submit.iloc[:,1:] > 50000, 50000, submit.iloc[:,1:])

differences = (submit_origin.iloc[:, 1:] != submit.iloc[:, 1:])
diff_list = differences.any(axis=1)
diff_list = diff_list.index[diff_list].tolist()

len(diff_list) # -1 : 3530

submit.to_csv('DLinear_after_process.csv', index=False)
